refactor: split App to Home
